Relocated openpyxl config file to configs folder
