+++ adaptive_hockey_federation/core/utils.py
@@ -3,7 +3,7 @@ from datetime import datetime
 from typing import Any, List, Optional
 
 from core.constants import AgeLimits, FileConstants, TimeFormat
-from core.settings.openpyxl_settings import (
+from core.config.openpyxl.settings import (
     ALIGNMENT_CENTER,
     HEADERS_BORDER,
     HEADERS_FILL,
